alot of things

+++ SimpleRPG/src/com/simplerpg/inventory/Weapon.java
@@ -2,14 +2,14 @@ package com.simplerpg.inventory;
 
 // A decorator is applied to a weapon to enhance/reduce its damage or add special effects.
 public class Weapon extends Item {
-    protected int baseDamage;
+    protected double baseDamage;
 
-    public Weapon(String name, int baseDamage) {
+    public Weapon(String name, double baseDamage) {
         super(name);
         this.baseDamage = baseDamage;
     }
 
-    public int getBaseDamage() {
+    public double getBaseDamage() {
         return baseDamage;
     }
